driver: fix proc_ops .owner removal and remaining issues
- Remove .owner = THIS_MODULE from all proc_ops structs (proc_ops
  has no owner field, unlike file_operations)
- Fix .release -> .proc_release and .llseek -> .proc_lseek in vic
- Add forward declaration of get_timestamp_us() before first use
- Fix var_28 type: unsigned long -> unsigned int for data_b2ee0 arg

+++ driver/tx_isp_tuning.c
@@ -31930,7 +31930,7 @@ static void tisp_set_sensor_integration_time(uint32_t time)
  * provide the full requested analog gain. */
 static uint32_t tisp_set_sensor_analog_gain(uint32_t requested_gain)
 {
-    unsigned long var_28;
+    unsigned int var_28;
     uint32_t log_result, gain_param, v0_2, final_gain;
 
     pr_debug("tisp_set_sensor_analog_gain: requested gain=0x%x\n", requested_gain);
